+++ samples/hello/2.server/Hello2Server.cs
@@ -21,9 +21,7 @@ namespace hello
         protected override void Setup()
         {
             EventFactory.Register<HelloReq>();
-
             Bind(new HelloResp(), Send);
-
             Listen(6789);
         }