+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Features:
 
-*TcpClient: removed synchronous connect interface
+- introduced link strategies to encapsulate pluggable functionalities
+- SessionBasedLink: removed session recovery support
+- added helloworld samples
+- *TcpClient: removed synchronous connect interface
 
 ## 0.9.19 (2018-01-26)